pyproject.toml
[project]
name = "open-shop"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-ckeditor-5>=0.2.18",
    "django-jazzmin>=3.0.1",
    "django-mptt>=0.18.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.11",
]
